requirements.txt
pandas
openpyxl
streamlit
altair
Pillow
treamlit_option_menu
